+++ src/main/java/org/polaris2023/relativity/nativeaccess/RcQuat.java
@@ -1,5 +1,5 @@
 package org.polaris2023.relativity.nativeaccess;
 
-public record RcQuat(float i, float j, float k, float w) {
-    public static final RcQuat IDENTITY = new RcQuat(0.0f, 0.0f, 0.0f, 1.0f);
+public record RcQuat(double i, double j, double k, double w) {
+    public static final RcQuat IDENTITY = new RcQuat(0.0, 0.0, 0.0, 1.0);
 }
